Added a few extra features, fixed ms to sec parsing. Made better, less redundant composables.

+++ app/src/main/java/com/example/trackpro/CalculationClasses/DragTimeCalculation.kt
@@ -34,7 +34,7 @@ class DragTimeCalculation(
             postProcessing()
         }
 
-        val sessionItems: List<SmoothedGPSData> = database.smoothedDataDao().getSmoothedGPSDataBySession(sessionid.toInt())
+        val sessionItems: List<SmoothedGPSData> = database.smoothedDataDao().getSmoothedGPSDataBySession(sessionid)
 
         var lastZeroTime: Long? = null
         var endTime: Long? = null
